Create LifeOnHitLocal.txt
This was created in the event that the life gain on hit suffix for both global and local get changed down the road.  For now it's not necessary to show the correct ilvl or affix, but this was just something I was going to do and forgot about.

LifeOnHitLocal is for weapons and LifeOnHit is for everything else.