Changed Life Gain on Hit
Changed Life Gain on Hit to accommodate both local and global values of the suffix.  If anything is changed down the road such as the local or global getting another higher ilvl affix, but they are not the same, this will already be in place to handle that.  It's just so happens they are exactly the same right now.

"Life gained for each Enemy hit by Attacks" (local = weapon) and "Life gained for each Enemiy hit by your Attacks" (global = gear)  are in fact different and both in the game.