Update MapList.txt
Added what octant each map is a part of in the new conquerers of the atlas expansion. This will help players determine which maps are a part of the octant they want to be exploring without toggling between the stash and the atlas multiple times. Unfortuneately these will probably need to be updated each league, so added maintence for each league beyond what they can turn into.